services/trade_ingestion/requirements.txt
websocket-client==1.7.0
psycopg2-binary==2.9.9
boto3==1.34.0
pyarrow==15.0.0
pandas==2.2.0
